requirements.txt
openmeteo-requests
numpy
pandas
retry-requests
requests-cache
asyncio
aiohttp
python-dotenv
click
requests
pyzmq